Fix local/html5fs/syncfs operations, sort of
Due to a typo, Javascript always opens TEMPORARY (=0) file system
because undefined value is interpreted as 0.  This is so frustrating.

W3C spec says it is an error to try to "copy any element to a path
occupied by a directory which is not empty."  So I guess I have no
choice but have to remove destination directory before copy to it.

But this is really bad because now every time a user uploads his C drive
to Google Drive, Chrome will do a full upload instead of an incremental
upload (remove C drive on Google Drive and then upload full contents).

If you wonder why I have to copy C drive back and forth between file
systems, it is because Native Client modules cannot open syncfs :(