Fix and use OpenGL in DOSBox/Pepper
I did not dig into the bottom of the problem, but basically it was that
Regal glNewList could not create an OpenGL display list (glIsList always
returned false).  To workaround this problem, I created a function that
replicated function calls of the display list.

I should revert this change once Regal team fixed this bug.